+++ Software_Requirement_Specification/Models/TaiLieu.cs
@@ -13,6 +13,7 @@ namespace Software_Requirement_Specification.Models
         //Khởi tạo 2:41 13/4/2022
         [Key]
         public int Id { get; set; }
+        public string TenMonHoc { get; set; }
         public MonHoc MonHocID { get; set; }
         public int SoTaiLieuChoDuyet { get; set; }
         public DateTime NgayGuiPheDuyet { get; set; }
