+++ Software_Requirement_Specification/appsettings.json
@@ -8,6 +8,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "Software_Requirement_SpecificationContext": "Data Source=DESKTOP-519BTCM\\SQLEXPRESS;Initial Catalog=E_Lybary;Integrated Security=True"
+    "Software_Requirement_SpecificationContext": "Data Source=DESKTOP-519BTCM\\SQLEXPRESS;Initial Catalog=E_Libery;Integrated Security=True"
   }
 }
